+++ CHANGELOG.md
@@ -3,12 +3,12 @@
 ## 3.3.0 (2021-11-11)
 Improvements:
 - Migrate QRIS (QR Codes) to XenInvoice
+- Add category on create invoice
 
 ## 3.2.0 (2021-11-02)
 Improvements:
 - Migrate Kredivo to XenInvoice
 - Data housekeeping
-- Add category on create invoice
 
 ## 3.1.0 (2021-10-03)
 Improvements: